Various fixes in setting globals in a script group
Bug: 25602504

1) Fixed setting 64-bit globals on a 32-bit platform.

The long values were casted to pointer-size integers (uintptr_t) or void*
in the runtime implementation. This caused the upper 32 bits to be lost
on 32-bit platforms. Fixed that by casting to int64_t instead.

2) Fixed setting Allocation values (either as arg or a global) in a script
group.

The size for a value in a script is signed, with -1 indicating that the
value is an Allocation. The sign has to be kept across different layers
of our stack.

3) Added a test for both native and support libraries for passing
floating point values to a script group, and for passing 64-bit values
(long or double) to a script group.

Change-Id: Ifff099a76be7707df7b67c388395f5a00f9cae66
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>